Extracted code from example usage
